Change Postgres config location
Previously, the postgres configuration file was added to the
/var/lib/postgresql/data directory which was overwritten by the
already-existing named data volume. When initializing a database,
the placement of the configuration file in /var/lib/postgresql/data
caused an error that the "PGDATA was not empty."

This commit fixes this issue by placing the configuration file in
a different location (/etc/postgresql) and then running the Dockerfile
entrypoint with an argument showing the location of the config file.