New: Private Collector Institutinos
This commit adds the ability to specify private collectors as a type of Institution.

The model constraints are set up so that you can either specify that an institution is a private collector (via a checkbox) OR they have a siglum. Both cannot be selected for the same record, but one of the two options must be selected for each record.

Institutions that do not have a siglum currently get the word "Private" prepended to the shelfmark.

All other fields are accessible for private collectors.

Alters the admin UI to make the distinction via the layout as well.