feat(source model): add temporary segment field to source
Adds the segment_m2m many-to-many field to the source model.
This additional field ensures that the data currently in the
segment field is not overwritten. Once data in the segment
field is transferred, this field can be renamed.

Changes to the source create and source edit forms make this
new field editable.

Refs: #1549