feat(chant create): use fixed nextchants API for suggested chants feature
With a fix to the Cantus Index json-nextchants endpoint, we can now
reduce the number of calls that need to be made to Cantus Index to
support the suggested chants feature of the Chant Create page. All
the data on the suggested chants that we need for the feature comes
now in the result of that call and does not require additional json-cid
calls.