New: Add views for institutions
This PR adds views for institutions, giving them each their own page on the Cantus site.

Special care was taken to get the source counts and source links correct for logged in / not logged in users and published / unpublished sources.

In addition external authorities are shown -- currently only RISM Online entries are in the external authorities, but others can be added. An entry for DIAMM as an external authority was also added so that links between Cantus and DIAMM are possible.