fix: maintain compatibility in external/export APIs
The addition of the institution model and the change from office to
service introduced changes in external APIs (the JSON CID export,  the JSON melody export, and the cached concordances command).
This commit rolls back those changes to maintain compatibility.  The modifications may be reintroduced at a later date if notice is given  to external users (in particular, Cantus Index).

Also simplifies the logic and adjusts the typing of functions related to the
JSON melody export and the AJAX melody list views to address #1638.

Makes relevant changes to associated tests.

Fixes additional type errors in views/api.py.

Refs #1641.