fix(source): make holding_institution optional for sources
This commit removes the requirement for a source to have a holding_institution.

Tests and forms are updated to account for this.

The `migrate_records` command is updated so that sources with un-parseable sigla that are not otherwise accounted for (by virtue of being private collections or prints) do not have institution records made.

Note: For the purposes of data migration, no constraint is added to sources to designate when no holding institution is necessary. After migration, these should be modified.

Refs: #1631