feat(edit chants): allow chant edit to save with no initial full text
Allow edits to chants that did not have an initial full text to be saved even if no full text is added.

Add check for uniqueness of folio-c_sequence to the chant edit page. This requires inclusion of the chant source in the chant edit form: the form, view, and test are edited to add the source to the form.

Refs: #1696.