Optimized chant search page queries
The use of the ONLY_FIELDS list can optimize loading of records, but if a record needs access to a field that isn't in the ONLY_FIELDS list, it will trigger another SQL query. The source__name was being accessed (to show the heading of the source) but it was not in the list, triggering another SQL query for each entry.

The addition of this value reduces the number of SQL queries from ~108 to ~8.