fix: remove unnecessary page_obj from UserSourceList context
Prior to 00cf05ffabff26195ecb9177b05f4eed78a6944b, we were doing some unnecessary work to paginate the main list of sources on the UserSourceList page, when the generic ListView would have taken care of this for us. This removes the last vestiges of this. Pagination is now managed by the built-in process.