fix: specify build config for celery-worker service
Docker seems to inconsistently recognize that it has built the image it needs for the `celery-worker` service in the `cdb-django` image, and sometimes goes looking for it on external repositories. Since adding the specific build configuration doesn't drain resources (as long as we are relying on the build cache), this new configuration should be more robust.