feat: Calculate and store proofreading statistics
This commit introduces a new model, `ProofreadingStats`, to track the progress of proofreading for each source. It includes:

- Model definition with fields for various proofreading metrics.
- A custom manager with a method to calculate and update these stats.
- A management command to trigger the calculation and update for all sources.
- Adds the ProofreadingStats model to the main_app models init file.

The stats include counts of chants needing proofreading for volpiano, manuscript full text, standardized spelling, and other fields, as well as overall progress. This information is used to provide a more granular view of the proofreading effort.